Add Style/CollectionMethods cop for e.g. map vs collect
Based on this comment: https://github.com/barsoom/auctionet/commit/95b79b013183e78505aa996d2631f692b508f3ba#r82038858

Most of these are strong conventions for us. reduce vs inject is less
so, but we have a lot more of the latter in Auctionet Core, and we might
as well be consistent.